✨Feat: input 컴퍼넌트 가이드 작성
